refactor: move CompletenessProof + foldCompletenessProof to csmt-core
