+++ mts.cabal
@@ -115,6 +115,7 @@ library csmt-core
   visibility:       public
   exposed-modules:
     CSMT.Core.CBOR
+    CSMT.Core.Completeness
     CSMT.Core.Exclusion
     CSMT.Core.Hash
     CSMT.Core.Proof
@@ -125,6 +126,7 @@ library csmt-core
     , bytestring  >=0.12 && <0.13
     , cborg       >=0.2  && <0.3
     , cereal      >=0.5  && <0.6
+    , containers  >=0.6  && <0.7
 
   hs-source-dirs:   lib/csmt-core/
   default-language: Haskell2010
